// client-clean-architecture/src/modules/fractal/presentation/types/stores/editor/visual-types.ts
export type CameraType = 'perspective' | 'orthographic';

export interface PerspectiveSettings {
    fov: number;
    near: number;
    far: number;
    zoom: number;
    focus: number;
    filmGauge: number;
    filmOffset: number;
    aspect: number;
    enableAutoFocus: boolean;
    autoFocusSpeed: number;
    bokehScale: number;
    maxBlur: number;
};

export interface OrthographicSettings {
    near: number;
    far: number;
    zoom: number;
};

export interface CameraSettingsState {
    type: CameraType;
    position: [number, number, number];
    up: [number, number, number];
    perspective: PerspectiveSettings;
    orthographic: OrthographicSettings;
};